# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shotstash REST API",
    "version": "0.1.0",
    "description": "REST routes of a Shotstash instance: auth, uploads, media, share pages, health and the processing worker contract (v1). Everything else (projects, files, sharing, discussion, jobs in the UI) is GraphQL at /api/graphql; see schema.graphql. Errors answer `{ code, message }` with a stable code. Until first-run setup is done every route except health, config and setup answers 503 SETUP_REQUIRED.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "The instance serving this document"
    }
  ],
...
  "paths": {
    "/api/graphql": {
      "get": {
        "operationId": "getApiGraphql",
        "summary": "Run a GraphQL query (GET)",
        "description": "GraphQL over GET for queries only (query and variables in the URL). The schema and every operation are documented in schema.graphql at the repository root; each root field declares its own auth (public or Bearer session).",
        "tags": [
          "GraphQL"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The GraphQL document."
            },
            "description": "The GraphQL document."
          },
          {
            "in": "query",
            "name": "variables",
            "required": false,
            "schema": {
              "type": "object",
              "additionalProperties": {},
              "description": "Operation variables."
            },
            "description": "Operation variables.",
            "style": "deepObject",
            "explode": true
          },
          {
            "in": "query",
            "name": "operationName",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Which operation of the document to run."
            },
            "description": "Which operation of the document to run."
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "GraphQL result (errors are inside the body)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphQLResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/SetupRequired"
          }
        },
        "x-shotstash-auth": "public"
      },
      "post": {
        "operationId": "postApiGraphql",
        "summary": "Run a GraphQL operation",
        "description": "The GraphQL endpoint for queries and mutations; subscriptions use graphql-ws on the same path. The schema and every operation are documented in schema.graphql at the repository root. Send the session as a Bearer token; each root field declares its own auth (public or session). The media cookie is never read here.",
        "tags": [
          "GraphQL"
        ],
        "parameters": [],
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphQLRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GraphQL result (errors are inside the body)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphQLResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/SetupRequired"
          }
        },
        "x-shotstash-auth": "public"
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "postApiV1AuthLogin",
        "summary": "Sign in",
        "description": "Email and password in and a Bearer token out; also sets the shotstash_session media cookie. Rate limited to 10 attempts per 15 minutes per IP and per email.",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST (email and password are required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_CREDENTIALS or another login refusal code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/SetupRequired"
          }
        },
        "x-shotstash-auth": "public"
      }
    },
...
    "/api/v1/pipeline/jobs/{id}/output": {
      "put": {
        "operationId": "putApiV1PipelineJobsByIdOutput",
        "summary": "Upload the job output",
        "description": "The raw bytes of the one output; streamed. Content-Type is its media type (video/mp4) and X-Output-Ext its extension (mp4). The size limit is SHOTSTASH_PIPELINE_MAX_OUTPUT_MB.",
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Job id (UUID)."
            },
            "description": "Job id (UUID)."
          },
          {
            "in": "header",
            "name": "X-Claim-Token",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Claim token from the claim answer."
            },
            "description": "Claim token from the claim answer."
          },
          {
            "in": "header",
            "name": "X-Output-Ext",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Extension of the output, such as `mp4`."
            },
            "description": "Extension of the output, such as `mp4`."
          }
        ],
        "security": [
          {
            "WorkerToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "contentMediaType": "application/octet-stream",
                "description": "Raw bytes, streamed."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_OUTPUT or CLAIM_TOKEN_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or revoked worker token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "CLAIM_STALE or JOB_TERMINAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "OUTPUT_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "STORAGE_UNAVAILABLE or SETUP_REQUIRED (first-run setup not done)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "x-shotstash-auth": "worker"
      }
    },
...
    "/api/v1/uploads/{sessionId}/parts/{partNumber}": {
      "put": {
        "operationId": "putApiV1UploadsBySessionIdPartsByPartNumber",
        "summary": "Upload one part",
        "description": "Streams one part of a resumable upload started with the initiateUpload mutation. The body is the raw bytes; Content-MD5 is required and checked. Sending a part again replaces it.",
        "tags": [
          "Uploads"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Upload session id (UUID) from the `initiateUpload` mutation."
            },
            "description": "Upload session id (UUID) from the `initiateUpload` mutation."
          },
          {
            "in": "path",
            "name": "partNumber",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Part number, from 1 to the session's `partCount`."
            },
            "description": "Part number, from 1 to the session's `partCount`."
          },
          {
            "in": "header",
            "name": "Content-MD5",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Base64 MD5 of the part body (RFC 1864); the server checks it."
            },
            "description": "Base64 MD5 of the part body (RFC 1864); the server checks it."
          }
        ],
        "security": [
          {
            "SessionToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "contentMediaType": "application/octet-stream",
                "description": "Raw bytes, streamed."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Part stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadPartResponse"
                }
              }
            }
          },
          "400": {
            "description": "PART_CHECKSUM_MISMATCH or PART_CHECKSUM_REQUIRED or PART_SIZE_MISMATCH or INVALID_PART_NUMBER",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN (another user session)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "UPLOAD_SESSION_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "UPLOAD_SESSION_CLOSED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "410": {
            "description": "UPLOAD_SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "STORAGE_UNAVAILABLE or SETUP_REQUIRED (first-run setup not done)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "x-shotstash-auth": "session"
      }
    },
...
  },
  "components": {
    "responses": {
      "SetupRequired": {
        "description": "SETUP_REQUIRED: first-run setup is not done yet. Every route except health, config and setup answers this until the first super admin exists.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, such as `NOT_FOUND` or `RATE_LIMITED`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable English message; never shown as is to end users."
          },
          "field": {
            "type": "string",
            "description": "The input field the error is about, when there is one."
          },
          "retryAfter": {
            "type": "number",
            "description": "Seconds to wait before trying again (429 answers only; also in `Retry-After`)."
          },
          "details": {
            "type": "object",
            "additionalProperties": {},
            "description": "Extra machine-readable details for some codes."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "GraphQLRequest": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "description": "The GraphQL document."
          },
          "variables": {
            "type": "object",
            "additionalProperties": {},
            "description": "Operation variables."
          },
          "operationName": {
            "type": "string",
            "description": "Which operation of the document to run."
          }
        },
        "required": [
          "query"
        ]
      },
      "GraphQLResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {},
            "description": "The result, shaped by the query."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {}
            },
            "description": "Errors; each carries `extensions.code`."
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email."
          },
          "password": {
            "type": "string",
            "description": "Account password."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Bearer token for `/api/graphql` and `/api/v1/*`. Keep it secret."
          },
          "expiresAt": {
            "type": "string",
            "description": "When the session expires unless used (ISO-8601 UTC)."
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "User id."
              },
              "email": {
                "type": "string",
                "description": "Account email."
              },
              "name": {
                "type": "string",
                "description": "Display name."
              },
              "role": {
                "type": "string",
                "enum": [
                  "SUPER_ADMIN",
                  "ADMIN",
                  "FIELD_CREW",
                  "EDITOR",
                  "VIEWER"
                ],
                "description": "Instance role."
              },
              "avatarUrl": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Relative URL of the avatar, or null."
              },
              "accountStatus": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "ACTIVE",
                  "REJECTED"
                ],
                "description": "PENDING accounts may sign in to see the waiting screen."
              },
              "onboardedAt": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "When onboarding was finished (ISO-8601 UTC), or null."
              }
            },
            "required": [
              "id",
              "email",
              "name",
              "role",
              "avatarUrl",
              "accountStatus",
              "onboardedAt"
            ],
            "description": "The signed-in account."
          }
        },
        "required": [
          "token",
          "expiresAt",
          "user"
        ]
      },
...
      "OutputResponse": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "string",
            "description": "Id of the stored output (it becomes a processed version on complete)."
          },
          "size": {
            "type": "number",
            "description": "Bytes stored."
          },
          "mimeType": {
            "type": "string",
            "description": "Media type of the output."
          }
        },
        "required": [
          "versionId",
          "size",
          "mimeType"
        ]
      },
...
      "UploadPartResponse": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "description": "The part that was stored."
          },
          "size": {
            "type": "number",
            "description": "Bytes stored for this part."
          },
          "confirmedParts": {
            "type": "number",
            "description": "How many parts of the session are stored so far (a count, not a list)."
          },
          "partCount": {
            "type": "number",
            "description": "Parts the session needs in total."
          }
        },
        "required": [
          "partNumber",
          "size",
          "confirmedParts",
          "partCount"
        ]
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "shotstash_session",
        "description": "The same session as an HttpOnly cookie scoped to /media (set by login and POST /api/v1/auth/cookie). Only /media routes read it."
      },
      "SessionToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from POST /api/v1/auth/login or the login mutation, sent as `Authorization: Bearer <token>`."
      },
      "ShareCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "shotstash_share_{slug}",
        "description": "Access to a PRIVATE share link, set by POST /s/{slug}/unlock and scoped to /s/{slug}. The real cookie name ends with the link slug. PUBLIC links need no cookie."
      },
      "WorkerBootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Worker-Bootstrap-Token",
        "description": "The shared WORKER_BOOTSTRAP_TOKEN of the instance; accepted by worker registration only."
      },
      "WorkerToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Worker-Token",
        "description": "Per-worker token returned once by worker registration."
      }
    }
  }
}